tests/requirements.txt
psycopg2>=2.6
django-reversion>=2
djangocms-text-ckeditor
djangocms-link
djangocms-helper>=0.9.2,<0.10
diff-match-patch==20110725.1
tox
coverage